Lowered timeout limits, made error messages less verbose
Every command has to be handled within 3 seconds so it doesn't time out. 
New timings might be too short. An alternative could be to make the output public, which would allow to delay the response.

Also reduced the error output to the exception name instead of the whole error message.